requirements.txt
fastapi
uvicorn
pywebview
requests
